requirements.txt
chromadb==1.0.12
langchain==0.3.25
langchain_chroma==0.2.4
langchain_openai==0.3.21
numpy==2.3.0
openai==1.84.0
pandas==2.3.0
python-dotenv==1.1.0
PyMuPDF==1.26.1
streamlit==1.44.1
